Implemented Resilience4j RateLimiter and tested using jmeter
Rate limiting is an imperative technique to prepare your API for scale and establish high availability and reliability of your service. But also, this technique comes with a whole bunch of different options of how to handle a detected limits surplus, or what type of requests you want to limit. You can simply decline this over limit request, or build a queue to execute them later or combine these two approaches in some way.

Co-Authored-By: Deendayal Kumawat <44318598+ddsha441981@users.noreply.github.com>